Move CA1017 to its new location.
