+++ PortFxCop/src/Microsoft.ApiDesignGuidelines.Analyzers/CSharp/Microsoft.ApiDesignGuidelines.CSharp.Analyzers.csproj
@@ -125,8 +125,6 @@
     <Compile Include="CSharpMarkAssembliesWithClsCompliant.Fixer.cs" />
     <Compile Include="CSharpMarkAssembliesWithAssemblyVersion.cs" />
     <Compile Include="CSharpMarkAssembliesWithAssemblyVersion.Fixer.cs" />
-    <Compile Include="CSharpMarkAssembliesWithComVisible.cs" />
-    <Compile Include="CSharpMarkAssembliesWithComVisible.Fixer.cs" />
     <Compile Include="CSharpMarkAttributesWithAttributeUsage.cs" />
     <Compile Include="CSharpMarkAttributesWithAttributeUsage.Fixer.cs" />
     <Compile Include="CSharpDefineAccessorsForAttributeArguments.cs" />
